requirements.txt
# Add project dependencies here
streamlit
langchain
langchain-community
langgraph
langchain-openai
python-dotenv
faiss-cpu
pypdf
